requirements.txt
pandas
numpy
sklearn
scipy
matplotlib
python-twitter
watson-developer-cloud
